[♻️refactor]: 기존 파일의 dummy데이터 삭제하고 sidebar 수정된 경로로 변경
