+++ src/pages/OrderList/OrderList.tsx
@@ -2,22 +2,11 @@ import { useQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
 
 import BreadcrumbAndTitle from '@/components/common/Breadcrumb/BreadcrumbAndTitle';
-import SideBar from '@/components/common/SideBar';
+import SideBar from '@/components/common/SideBar/SideBar';
 import Spinner from '@/components/common/Spinner/Spinner';
 import { getOrderList } from '@/services/orderList/getOrderList';
 
 import OrderData from './OrderData';
-
-// ! 공통 컴포넌트 - SideBar 변경 후 더미데이터 제거 예정
-const DUMMY_SIDEBAR_DATA = [
-  { name: '프로필', url: '/profile' },
-  { name: '회원 정보 수정', url: '/profile/edit' },
-  { name: '로그아웃', url: '/logout' },
-  { name: '프로필', url: '/profile' },
-  { name: '회원 정보 수정', url: '/profile/edit' },
-  { name: '로그아웃', url: '/logout' },
-  { name: '프로필', url: '/profile' },
-];
 
 export default function OrderList() {
   const { data: orderList, isLoading } = useQuery({
@@ -53,7 +42,7 @@ export default function OrderList() {
           )}
         </main>
         <aside className="w-[12rem] whitespace-nowrap sticky top-[80px] h-fit">
-          <SideBar menuItems={DUMMY_SIDEBAR_DATA} />
+          <SideBar />
         </aside>
       </div>
     </>
